Revert "Filters doctors"
