+++ lib/feature/home/home_screen.dart
@@ -1,12 +1,15 @@
+import 'dart:developer';
 
 import 'package:daweny/core/helpers/spacing.dart';
+import 'package:daweny/feature/home/logic/cubit/home_cubit_cubit.dart';
 
-import 'package:daweny/feature/home/ui/widget/specilization/specilization_bloc_builder.dart';
-import 'package:daweny/feature/home/ui/widget/doctors/doctors_bloc_builder.dart';
+import 'package:daweny/feature/home/ui/widget/doctor_specility_listview.dart';
 import 'package:daweny/feature/home/ui/widget/doctors_blue_container.dart';
 import 'package:daweny/feature/home/ui/widget/home_appbar.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
+import 'ui/widget/recommendations_doctors.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
@@ -23,9 +26,42 @@ class HomeScreen extends StatelessWidget {
               const HomeAppbar(),
               const DoctorsBlueContainer(),
               verticalSpace(15),
-             const SpecilizationAndBlocBuilder(),
-             verticalSpace(8),
-             const DoctorsBlocBuilder()
+              BlocBuilder<HomeCubit, HomeState>(
+                buildWhen: (previous, current) =>
+                    current is Loading || current is Error || current is Loaded,
+                builder: (context, state) {
+                  return state.maybeWhen(
+                    orElse: () {
+                      return SizedBox();
+                    },
+                    loaded: (specilzationsModel) {
+                      var specializationDataList =
+                          specilzationsModel.specializationDataList;
+                      return Expanded(
+                        child: Column(
+                          children: [
+                            DoctorSpecilityListView(
+                                specializationDataList:
+                                    specializationDataList ?? []),
+                                    verticalSpace( 8),
+                            DoctorsListView(
+                              doctors: specializationDataList![0].doctors,
+                            ),
+                          ],
+                        ),
+                      );
+                    },
+                    loading: () {
+                      return const Center(child: CircularProgressIndicator());
+                    },
+                    error: (error) {
+                      log("Error1 $error");
+                      String? errorMessage = error.apiErrorModel.message;
+                      return Center(child: Text("Error: $errorMessage"));
+                    },
+                  );
+                },
+              ),
             ],
           ),
         ),
@@ -33,4 +69,3 @@ class HomeScreen extends StatelessWidget {
     );
   }
 }
-
