--- lib/feature/home/ui/widget/specilization/specility_shimmer_loading.dart
@@ -1,50 +1,0 @@
-import 'package:daweny/core/helpers/spacing.dart';
-import 'package:daweny/core/thems/colors/colors.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:shimmer/shimmer.dart';
-
-class SpecialityShimmerLoading extends StatelessWidget {
-  const SpecialityShimmerLoading({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: 100.h,
-      child: ListView.builder(
-        scrollDirection: Axis.horizontal,
-        itemCount: 8,
-        itemBuilder: (context, index) {
-          return Padding(
-            padding: EdgeInsetsDirectional.only(start: index == 0 ? 0 : 24.w),
-            child: Column(
-              children: [
-                Shimmer.fromColors(
-                  baseColor: MangerColors.gray,
-                  highlightColor: Colors.white,
-                  child: const CircleAvatar(
-                    radius: 28,
-                    backgroundColor: Colors.white,
-                  ),
-                ),
-                verticalSpace(14),
-                Shimmer.fromColors(
-                  baseColor: MangerColors.gray,
-                  highlightColor: Colors.white,
-                  child: Container(
-                    height: 14.h,
-                    width: 50.w,
-                    decoration: BoxDecoration(
-                      color:  MangerColors.gray,
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          );
-        },
-      ),
-    );
-  }
-}
