+++ pubspec.yaml
@@ -49,13 +49,11 @@ dependencies:
   flutter_native_splash: ^2.4.3
   easy_localization: ^3.0.0
   flutter_screenutil: ^5.9.3
-  flutter_svg: ^2.0.17
+  flutter_svg: ^2.0.16
   go_router: ^14.6.1
   bloc: ^8.1.4
   shared_preferences: ^2.5.1
   flutter_secure_storage: ^9.0.0
-  shimmer: ^3.0.0
-  cached_network_image: ^3.4.1
 
   
   
@@ -86,7 +84,6 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
-    - assets/svgs/specilty.svg
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
